fixed coding guidelines and added a lot of testcode

+++ MedianQuicksort.h
@@ -3,13 +3,15 @@
 size_t pivotPartition(std::vector<size_t>& values, size_t left, size_t right) {
     size_t pivotIndex = left + (right - left) / 2;
     size_t pivotValue = values[pivotIndex];
-    size_t i = left;
-    size_t j = right;
+    int i = left;
+    int j = right;
     while (i <= j) {
-        while (values[i] < pivotValue) {
+        while (values[i] < pivotValue)
+        {
             i++;
         }
-        while (values[j] > pivotValue) {
+        while (values[j] > pivotValue)
+        {
             j--;
         }
         if (i <= j) {
@@ -21,7 +23,8 @@ size_t pivotPartition(std::vector<size_t>& values, size_t left, size_t right) {
     return i;
 }
 
-void quicksort(std::vector<size_t>& values, size_t left, size_t right) {
+void quicksort(std::vector<size_t>& values, size_t left, size_t right)
+{
     if (left < right) {
         size_t pivotIndex = pivotPartition(values, left, right);
         quicksort(values, left, pivotIndex - 1);
@@ -29,7 +32,8 @@ void quicksort(std::vector<size_t>& values, size_t left, size_t right) {
     }
 }
 
-size_t getQuicksortMedian(std::vector<size_t> values, size_t i) {
+size_t getQuicksortMedian(std::vector<size_t> values, size_t i)
+{
     //std::qsort(numbers); // only takes array param -> custom implementation with vector
     quicksort(values, 0, values.size() - 1);
     return values[i];
